+++ R/modelStudio.R
@@ -37,6 +37,7 @@
 #' @param N Number of observations used for the calculation of PD and AD. Default is \code{300}.
 #'  See \href{https://modelstudio.drwhy.ai/articles/ms-perks-features.html#more-calculations-means-more-time}{\bold{vignette}}
 #' @param N_fi Number of observations used for the calculation of FI. Default is \code{10*N}.
+#' @param N_sv Number of observations used for the calculation of SV. Default is \code{3*N}.
 #' @param B Number of permutation rounds used for calculation of SV. Default is \code{10}.
 #'  See \href{https://modelstudio.drwhy.ai/articles/ms-perks-features.html#more-calculations-means-more-time}{\bold{vignette}}
 #' @param B_fi Number of permutation rounds used for calculation of FI. Default is \code{B}.
@@ -178,6 +179,7 @@ modelStudio.explainer <- function(explainer,
                                   max_features = 10,
                                   N = 300,
                                   N_fi = N*10,
+                                  N_sv = N*3,
                                   B = 10,
                                   B_fi = B,
                                   eda = TRUE,
@@ -265,6 +267,12 @@ modelStudio.explainer <- function(explainer,
   obs_count <- dim(new_observation)[1]
   obs_data <- new_observation
   obs_list <- list()
+  
+  if (!is.null(N_sv) && N_sv < nrow(data)) {
+    data_sv <- data[sample(1:nrow(data), N_sv),, drop = FALSE]
+  } else {
+    data_sv <- data
+  }
 
   ## later update progress bar after all explanation functions
   if (show_info) {
@@ -374,7 +382,7 @@ modelStudio.explainer <- function(explainer,
         paste0("iBreakDown::local_attributions (", i, ")      "), show_info, pb, 2)
       sv <- calculate(
         iBreakDown::shap(
-          model, data, predict_function, new_observation, label = label, B = B),
+          model, data_sv, predict_function, new_observation, label = label, B = B),
         paste0("iBreakDown::shap (", i, ")                    "), show_info, pb, 3*B)
       cp <- calculate(
         ingredients::ceteris_paribus(
@@ -414,7 +422,7 @@ modelStudio.explainer <- function(explainer,
         paste0("iBreakDown::local_attributions (", i, ")      "), show_info, pb, 2)
       sv <- calculate(
         iBreakDown::shap(
-          model, data, predict_function, new_observation, label = label, B = B),
+          model, data_sv, predict_function, new_observation, label = label, B = B),
         paste0("iBreakDown::shap (", i, ")                    "), show_info, pb, 3*B)
       cp <- calculate(
         ingredients::ceteris_paribus(